pyproject.toml
[project]
name = "medicc-todo-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Dale Villa",email = "arvilla3483@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-rest-knox (>=5.0.2,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "dry-rest-permissions (>=0.1.10,<0.2.0)",
    "django-phonenumber-field[phonenumbers] (>=8.0.0,<9.0.0)",
    "django-money (>=3.5.3,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.2"}
commitizen = "^4.1.0"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings.test"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
